requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
overturemaps>=0.4.0
geopandas>=0.14.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
